proxy-service/requirements.txt
mitmproxy==10.3.0
httpx==0.27.2
PyJWT==2.10.1
python-dotenv==1.0.1
